+++ composeApp/src/androidMain/kotlin/com/whatever/caramel/Previews.kt
@@ -3,8 +3,8 @@ package com.whatever.caramel
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import com.whatever.caramel.feature.splash.SplashState
-import com.whatever.caramel.feature.splash.SplashScreen
+import com.whatever.caramel.feat.splash.SplashState
+import com.whatever.caramel.feat.splash.SplashScreen
 
 @Preview
 @Composable
