+++ composeApp/src/commonMain/kotlin/com/whatever/caramel/di/Modules.kt
@@ -3,7 +3,7 @@ package com.whatever.caramel.di
 import com.whatever.caramel.core.data.HttpClientFactory
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.core.module.Module
-import com.whatever.caramel.feature.splash.SplashViewModel
+import com.whatever.caramel.feat.splash.SplashViewModel
 import org.koin.dsl.module
 
 // expect는 platform별로 패키지까지 동일하게 사용해야함
