+++ composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/login/Navigator.kt
@@ -1,0 +1,10 @@
+package com.whatever.caramel.feat.login
+
+import cafe.adriel.voyager.core.registry.screenModule
+import com.whatever.caramel.app.Route
+
+val loginNavigator = screenModule {
+    register<Route.Login> {
+        LoginScreenRoot()
+    }
+}
