+++ composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/main/home/HomeTab.kt
@@ -1,0 +1,35 @@
+package com.whatever.caramel.feat.main.home
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Home
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.graphics.vector.rememberVectorPainter
+import app.composeapp.generated.resources.Res
+import app.composeapp.generated.resources.test_tab_item1
+import cafe.adriel.voyager.navigator.Navigator
+import cafe.adriel.voyager.navigator.tab.Tab
+import cafe.adriel.voyager.navigator.tab.TabOptions
+import org.jetbrains.compose.resources.stringResource
+
+object HomeTab : Tab {
+    override val options: TabOptions
+        @Composable
+        get(){
+            val title = stringResource(Res.string.test_tab_item1)
+            val icon = rememberVectorPainter(Icons.Default.Home)
+
+            return remember {
+                TabOptions(
+                    index = 0u,
+                    title = title,
+                    icon = icon
+                )
+            }
+        }
+
+    @Composable
+    override fun Content() {
+        Navigator(HomeScreenRoot())
+    }
+}
