+++ composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/onboarding/Navigator.kt
@@ -1,0 +1,10 @@
+package com.whatever.caramel.feat.onboarding
+
+import cafe.adriel.voyager.core.registry.screenModule
+import com.whatever.caramel.app.Route
+
+val onboardingNavigator = screenModule {
+    register<Route.Onboarding> {
+        OnboardingScreenRoot()
+    }
+}
