+++ composeApp/src/commonMain/kotlin/com/whatever/caramel/feat/setting/SettingScreen.kt
@@ -1,0 +1,29 @@
+package com.whatever.caramel.feat.setting
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import cafe.adriel.voyager.core.screen.Screen
+
+class SettingScreenRoot : Screen {
+    @Composable
+    override fun Content() {
+        SettingScreen()
+    }
+}
+
+@Composable
+fun SettingScreen(modifier: Modifier = Modifier) {
+    Column(
+        modifier = modifier.fillMaxSize(),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Text("Setting")
+    }
+}
+
