--- composeApp/src/commonMain/kotlin/com/whatever/caramel/feature/login/Navigator.kt
@@ -1,10 +1,0 @@
-package com.whatever.caramel.feature.login
-
-import cafe.adriel.voyager.core.registry.screenModule
-import com.whatever.caramel.app.Route
-
-val loginNavigator = screenModule {
-    register<Route.Login> {
-        LoginScreenRoot()
-    }
-}
