--- composeApp/src/commonMain/kotlin/com/whatever/caramel/feature/main/Navigator.kt
@@ -1,10 +1,0 @@
-package com.whatever.caramel.feature.main
-
-import cafe.adriel.voyager.core.registry.screenModule
-import com.whatever.caramel.app.Route
-
-val mainNavigator = screenModule {
-    register<Route.Main> {
-        MainScreenRoot()
-    }
-}
