--- composeApp/src/commonMain/kotlin/com/whatever/caramel/feature/main/home/HomeScreen.kt
@@ -1,40 +1,0 @@
-package com.whatever.caramel.feature.main.home
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.unit.sp
-import cafe.adriel.voyager.core.screen.Screen
-
-class HomeScreenRoot : Screen {
-    @Composable
-    override fun Content() {
-        HomeScreen()
-    }
-}
-
-@Composable
-fun HomeScreen(modifier: Modifier = Modifier) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(
-            text = "Home",
-            style = TextStyle(
-                color = Color.Black,
-                fontSize = 24.sp
-            ),
-            modifier = Modifier
-        )
-    }
-}
